Separate indexentries manager to sphinx.environment.manager.indexentries
